Remove cpp_*_id surrogate keys: all universe/market/blueprint models now use the EVE game ID directly as the primary key.

+++ evebs/models/tables/associations.py
@@ -9,7 +9,7 @@ eve_items_users = db.Table('eve_items_users',
 trade_hubs_users = db.Table('trade_hubs_users',
     db.Column('id', db.Integer, primary_key=True),
     db.Column('user_id', db.Integer, db.ForeignKey('users.id')),
-    db.Column('universe_system_id', db.BigInteger, db.ForeignKey('universe_systems.id')),
+    db.Column('universe_system_id', db.Integer, db.ForeignKey('universe_systems.id')),
 )
 
 user_blueprints = db.Table('user_blueprints',
