+++ evebs/models/tables/weekly_price_detail.py
@@ -12,7 +12,7 @@ class WeeklyPriceDetail(db.Model):
 
     id = db.Column(db.BigInteger, primary_key=True)
     eve_item_id = db.Column(db.BigInteger, db.ForeignKey('eve_items.id'), nullable=False)
-    universe_system_id = db.Column(db.BigInteger, db.ForeignKey('universe_systems.id'), nullable=False)
+    universe_system_id = db.Column(db.Integer, db.ForeignKey('universe_systems.id'), nullable=False)
     day = db.Column(db.Date, nullable=False)
     volume = db.Column(db.Float, nullable=False)
     weighted_avg_price = db.Column(db.Float, nullable=False)
